requirements.txt
anthropic[mcp]>=0.40.0
mcp>=1.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
python-telegram-bot>=20.0
